Module clarinet::test::ast

Structs

BundleHook

DiagnosticBuffer
A buffer for collecting diagnostic messages from the AST parser.

EmitOptions
Options which can be adjusted when transpiling a module.

ErrorBuffer
A buffer for collecting errors from the AST parser.

LexedItem

Location

ParsedModule
A logical structure to hold the value of a parsed module for further processing.

Enums

ImportsNotUsedAsValues

TokenOrComment

Statics

TARGET

Functions

flatten_comments

get_es_config

get_syntax

get_ts_config

lex

parse
For a given specifier, source, and media type, parse the source of the module and return a representation which can be further processed.

parse_with_source_map

strip_config_from_emit_options

transpile_module
A low level function which transpiles a source module into an swc SourceFile.