Module clarinet::test::file_fetcher

Structs

File
A structure representing a source file.

FileCache
Simple struct implementing in-process caching to prevent multiple fs reads/net fetches for same file.

FileFetcher
A structure for resolving, fetching and caching source files.

Enums

CacheSetting
Indicates how cached source files should be handled.

Constants

SUPPORTED_SCHEMES

Statics

DENO_AUTH_TOKENS

Functions

fetch_local
Fetch a source file from the local file system.

get_source_from_bytes
Given a vector of bytes and optionally a charset, decode the bytes to a string.

get_source_from_data_url

get_validated_scheme
Return a validated scheme for a given module specifier.

map_content_type
Resolve a media type and optionally the charset from a module specifier and the value of a content type header.

map_js_like_extension
Used to augment media types by using the path part of a module specifier to resolve to a more accurate media type.

strip_shebang
Remove shebangs from the start of source code strings