Function clarinet::test::file_watcher::watch_func_with_module_resolution

pub async fn watch_func_with_module_resolution<F, G, T>(
    module_resolver: F, 
    operation: G, 
    job_name: &str
) -> Result<(), AnyError> where
    F: Fn() -> Pin<Box<dyn Future<Output = ModuleResolutionResult<T>>>>,
    G: Fn(T) -> Pin<Box<dyn Future<Output = Result<(), AnyError>>>>,
    T: Clone, 

This function adds watcher functionality to subcommands like run or bundle. The difference from watch_func is that this does depend on ModuleGraph.

module_resolver is used for both resolving file paths to be watched at every restarting of the watcher and building ModuleGraph or ModuleSpecifier which will then be passed to operation.

operation is the actual operation we want to run every time the watcher detects file changes. For example, in the case where we would like to bundle, then operation would have the logic for it like doing bundle with the help of ModuleGraph.

job_name is just used for printing watcher status to terminal.

Note that the watcher will try to continue watching files using the previously resolved data if module_resolver fails at some point, which means the watcher won't work at all if module_resolver fails at the first attempt.