Module clarinet::test::fmt_errors

This mod provides DenoError to unify errors across Deno.

Structs

PrettyJsError
Wrapper around deno_core::JsError which provides colorful string representation.

Constants

SOURCE_ABBREV_THRESHOLD

Functions

cyan

default_color

format_frame

format_location

format_maybe_source_line
Take an optional source line and associated information to format it into a pretty printed version of that line.

format_stack

italic_bold

yellow