Struct clarinet::test::import_map::ImportMap

pub struct ImportMap {
    base_url: String,
    imports: IndexMap<String, Option<Url>>,
    scopes: IndexMap<String, IndexMap<String, Option<Url>>>,
}

Fields
base_url: Stringimports: IndexMap<String, Option<Url>>scopes: IndexMap<String, IndexMap<String, Option<Url>>>
Implementations

impl ImportMap

pub fn from_json(
base_url: &str,
json_string: &str
) -> Result<Self, ImportMapError>

fn try_url_like_specifier(specifier: &str, base: &str) -> Option<Url>

fn normalize_specifier_key(
specifier_key: &str,
base_url: &str,
diagnostics: &mut Vec<String>
) -> Option<String>

Parse provided key as import map specifier.

Specifiers must be valid URLs (eg. “https://deno.land/x/std/testing/asserts.ts”) or “bare” specifiers (eg. “moment”).

fn parse_specifier_map(
json_map: &Map<String, Value>,
base_url: &str,
diagnostics: &mut Vec<String>
) -> IndexMap<String, Option<Url>>

Convert provided JSON map to valid SpecifierMap.

From specification:

order of iteration must be retained

SpecifierMap’s keys are sorted in longest and alphabetic order

fn parse_scope_map(
scope_map: &Map<String, Value>,
base_url: &str,
diagnostics: &mut Vec<String>
) -> Result<IndexMap<String, IndexMap<String, Option<Url>>>, ImportMapError>

Convert provided JSON map to valid ScopeMap.

From specification:

order of iteration must be retained

ScopeMap’s keys are sorted in longest and alphabetic order

fn resolve_scopes_match(
scopes: &IndexMap<String, IndexMap<String, Option<Url>>>,
normalized_specifier: &str,
as_url: Option<&Url>,
referrer: &str
) -> Result<Option<Url>, ImportMapError>

fn resolve_imports_match(
specifier_map: &IndexMap<String, Option<Url>>,
normalized_specifier: &str,
as_url: Option<&Url>
) -> Result<Option<Url>, ImportMapError>

pub fn resolve(
&self,
specifier: &str,
referrer: &str
) -> Result<Option<Url>, ImportMapError>