Module clarinet::test::info

Structs

ModuleGraphInfo

ModuleGraphInfoDep

ModuleGraphInfoMod

ModuleInfoMapItem
An entry in the ModuleInfoMap the provides the size of the module and a vector of its dependencies, which should also be available as entries in the map.

Constants

CHILD_DEPS_CONNECTOR

CHILD_NO_DEPS_CONNECTOR

EMPTY_CONNECTOR

LAST_SIBLING_CONNECTOR

SIBLING_CONNECTOR

VERTICAL_CONNECTOR

Functions

human_size
A function that converts a float to a string the represents a human readable version of that number.