Struct clarinet::test::module_graph::Graph

pub struct Graph {
    handler: Arc<Mutex<dyn SpecifierHandler>>,
    maybe_tsbuildinfo: Option<String>,
    modules: HashMap<ModuleSpecifier, ModuleSlot>,
    redirects: HashMap<ModuleSpecifier, ModuleSpecifier>,
    roots: Vec<ModuleSpecifier>,
    roots_dynamic: bool,
    maybe_lockfile: Option<Arc<Mutex<Lockfile>>>,
}

A dependency graph of modules, were the modules that have been inserted via the builder will be loaded into the graph. Also provides an interface to be able to manipulate and handle the graph.

Fields
handler: Arc<Mutex<dyn SpecifierHandler>>

A reference to the specifier handler that will retrieve and cache modules for the graph.

maybe_tsbuildinfo: Option<String>

Optional TypeScript build info that will be passed to tsc if tsc is invoked.

modules: HashMap<ModuleSpecifier, ModuleSlot>

The modules that are part of the graph.

redirects: HashMap<ModuleSpecifier, ModuleSpecifier>

A map of redirects, where a module specifier is redirected to another module specifier by the handler. All modules references should be resolved internally via this, before attempting to access the module via the handler, to make sure the correct modules is being dealt with.

roots: Vec<ModuleSpecifier>

The module specifiers that have been uniquely added to the graph, which does not include any transient dependencies.

roots_dynamic: bool

If all of the root modules are dynamically imported, then this is true. This is used to ensure correct --reload behavior, where subsequent calls to a module graph where the emit is already valid do not cause the graph to re-emit.

maybe_lockfile: Option<Arc<Mutex<Lockfile>>>
Implementations

impl Graph

pub fn new(
handler: Arc<Mutex<dyn SpecifierHandler>>,
maybe_lockfile: Option<Arc<Mutex<Lockfile>>>
) -> Self

Create a new instance of a graph, ready to have modules loaded it.

The argument handler is an instance of a structure that implements the SpecifierHandler trait.

pub fn bundle(
&self,
options: BundleOptions
) -> Result<(String, Stats, Option<IgnoredCompilerOptions>), AnyError>

Transform the module graph into a single JavaScript module which is returned as a String in the result.

pub fn check(self, options: CheckOptions) -> Result<ResultInfo, AnyError>

Type check the module graph, corresponding to the options provided.

pub fn emit(
self,
options: EmitOptions
) -> Result<(HashMap<String, String>, ResultInfo), AnyError>

Emit the module graph in a specific format. This is specifically designed to be an “all-in-one” API for access by the runtime, allowing both emitting single modules as well as bundles, using Deno module resolution or supplied sources.

fn emit_bundle(
&self,
specifier: &ModuleSpecifier,
emit_options: &EmitOptions,
bundle_type: &BundleType
) -> Result<String, AnyError>

Shared between bundle() and emit().

fn flush(&mut self) -> Result<(), AnyError>

Update the handler with any modules that are marked as dirty and update any build info if present.

pub fn get_loadable_modules(
&self
) -> HashMap<ModuleSpecifier, Result<ModuleSource, AnyError>>

Retrieve a map that contains a representation of each module in the graph which can be used to provide code to a module loader without holding all the state to be able to operate on the graph.

pub fn get_media_type(&self, specifier: &ModuleSpecifier) -> Option<MediaType>

fn get_module(&self, specifier: &ModuleSpecifier) -> &ModuleSlot

fn get_module_mut(
&mut self,
specifier: &ModuleSpecifier
) -> Option<&mut ModuleSlot>

pub fn get_modules(&self) -> Vec<ModuleSpecifier>

Consume graph and return list of all module specifiers contained in the graph.

fn get_root_names(
&self,
include_emittable: bool
) -> Result<Vec<(ModuleSpecifier, MediaType)>, AnyError>

Transform self.roots into something that works for tsc, because tsc doesn’t like root names without extensions that match its expectations, nor does it have any concept of redirection, so we have to resolve all that upfront before feeding it to tsc. In addition, if checkJs is not true, we should pass all emittable files in as the roots, so that tsc type checks them and potentially emits them.

pub fn get_source(&self, specifier: &ModuleSpecifier) -> Option<String>

Get the source for a given module specifier. If the module is not part of the graph, the result will be None.

pub fn info(&self) -> Result<ModuleGraphInfo, AnyError>

Return a structure which provides information about the module graph and the relationship of the modules in the graph. This structure is used to provide information for the info subcommand.

fn is_emit_valid(&self, config: &TsConfig) -> bool

Determines if all of the modules in the graph that require an emit have a valid emit. Returns true if all the modules have a valid emit, otherwise false.

pub fn lock(&self)

Verify the subresource integrity of the graph based upon the optional lockfile, updating the lockfile with any missing resources. This will error if any of the resources do not match their lock status.

fn needs_emit(&self, config: &TsConfig) -> bool

Determines if any of the modules in the graph are required to be emitted. This is similar to emit_valid() except that the actual emit isn’t checked to determine if it is valid.

pub fn resolve(
&self,
specifier: &str,
referrer: &ModuleSpecifier,
prefer_types: bool
) -> Result<ModuleSpecifier, AnyError>

Given a string specifier and a referring module specifier, provide the resulting module specifier and media type for the module that is part of the graph.

Arguments

specifier - The string form of the module specifier that needs to be resolved.

referrer - The referring ModuleSpecifier.

prefer_types - When resolving to a module specifier, determine if a type dependency is preferred over a code dependency. This is set to true when resolving module names for tsc as it needs the type dependency over the code, while other consumers do not handle type only dependencies.

fn resolve_specifier<'a>(
&'a self,
specifier: &'a ModuleSpecifier
) -> &'a ModuleSpecifier

Takes a module specifier and returns the “final” specifier, accounting for any redirects that may have occurred.

pub fn transpile(
&mut self,
options: TranspileOptions
) -> Result<ResultInfo, AnyError>

Transpile (only transform) the graph, updating any emitted modules with the specifier handler. The result contains any performance stats from the compiler and optionally any user provided configuration compiler options that were ignored.

Arguments

options - A structure of options which impact how the code is transpiled.

fn validate(&self) -> Result<(), AnyError>

Validate that the module graph is “valid” in that there are not module slots that have errorred that should be available to be able to statically analyze. In certain situations, we can spin up tsc with an “invalid” graph.