Struct clarinet::test::program_state::ProgramState

pub struct ProgramState {
    pub flags: Flags,
    pub dir: DenoDir,
    pub coverage_dir: Option<String>,
    pub file_fetcher: FileFetcher,
    pub modules: Arc<Mutex<HashMap<ModuleSpecifier, Result<ModuleSource, AnyError>>>>,
    pub lockfile: Option<Arc<Mutex<Lockfile>>>,
    pub maybe_import_map: Option<ImportMap>,
    pub maybe_inspector_server: Option<Arc<InspectorServer>>,
    pub ca_data: Option<Vec<u8>>,
}

This structure represents state of single “deno” program.

It is shared by all created workers (thus V8 isolates).

Fields
flags: Flags

Flags parsed from argv contents.

dir: DenoDircoverage_dir: Option<String>file_fetcher: FileFetchermodules: Arc<Mutex<HashMap<ModuleSpecifier, Result<ModuleSource, AnyError>>>>lockfile: Option<Arc<Mutex<Lockfile>>>maybe_import_map: Option<ImportMap>maybe_inspector_server: Option<Arc<InspectorServer>>ca_data: Option<Vec<u8>>
Implementations

impl ProgramState

pub async fn build(flags: Flags) -> Result<Arc<Self>, AnyError>

pub async fn prepare_module_load(
self: &Arc<Self>,
specifier: ModuleSpecifier,
lib: TypeLib,
runtime_permissions: Permissions,
is_dynamic: bool,
maybe_import_map: Option<ImportMap>
) -> Result<(), AnyError>

This function is called when new module load is initialized by the JsRuntime. Its resposibility is to collect all dependencies and if it is required then also perform TS typecheck and traspilation.

pub fn load(
&self,
specifier: ModuleSpecifier,
maybe_referrer: Option<ModuleSpecifier>
) -> Result<ModuleSource, AnyError>

fn get_emit(&self, url: &Url) -> Option<(Vec<u8>, Option<Vec<u8>>)>