Module clarinet::test::specifier_handler

Structs

CachedModule

CompiledFileMetadata
A representation of meta data for a compiled file.

Dependency

FetchHandler
An implementation of the SpecifierHandler trait that integrates with the existing file_fetcher interface, which will eventually be refactored to align it more to the SpecifierHandler trait.

MemoryHandler

Enums

Emit
An enum to own the a specific emit.

HandlerError
A group of errors that represent errors that can occur with an an implementation of SpecifierHandler.

Traits

SpecifierHandler

Type Definitions

DependencyMap

FetchFuture