Module clarinet::test::tsc

Structs

CreateHashArgs

EmitArgs

EmittedFile

LoadArgs

Request
A structure representing a request to be sent to the tsc runtime.

ResolveArgs

RespondArgs

Response

STATIC_ASSETS
Contains static assets that are not preloaded in the compiler snapshot.

State

Statics

COMPILER_SNAPSHOT

Functions

compiler_snapshot

create_hash

emit

exec
Execute a request on the supplied snapshot, returning a response which contains information, like any emitted files, diagnostics, statistics and optionally an updated TypeScript build info.

get_asset
Retrieve a static asset that are included in the binary.

get_maybe_hash

get_tsc_media_type
tsc only supports .ts, .tsx, .d.ts, .js, or .jsx as root modules and so we have to detect the apparent media type based on extensions it supports.

hash_data_url

load

op

resolve

respond