Module clarinet::test::tsc_config

Structs

EmitConfigOptions
The transpile options that are significant out of a user provided tsconfig file, that we want to deserialize out of the final config for a transpile.

IgnoredCompilerOptions
A structure that represents a set of options that were ignored and the path those options came from.

TSConfigJson

TsConfig
A structure for managing the configuration of TypeScript

Constants

IGNORED_COMPILER_OPTIONS
A static slice of all the compiler options that should be ignored that either have no effect on the compilation or would cause the emit to not work in Deno.

IGNORED_RUNTIME_COMPILER_OPTIONS

Functions

json_merge
A function that works like JavaScript’s Object.assign().

jsonc_to_serde
Convert a jsonc libraries JsonValue to a serde Value.

parse_compiler_options

parse_config
Take a string of JSONC, parse it and return a serde Value of the text. The result also contains any options that were ignored.